requirements.txt
opencv-python>=4.8.0
onvif-zeep>=0.2.12
zeep>=4.2.1

